Expose an internal constructor of FreeAbelianGroup from .Internal module
